feat(app): Add infinite scroll on mobile.
This was quite hard because when you switch from mobile to desktop or viceversa the data mos modified.
To reach the solution I create an object with id by key in which we will add all the pets (like a redux state).

Also, in the service, we got a getPetList (like a redux selector) which search on this state an retrieve the neccessary information acording to some params.

After we use getPetList In the main component to render the data according to the type of device