feat(app): Implemented getPetOfTheDay() in PetsService
We check if already exist (using LocalStorage) and if exist we check that we are in the same day.
It fails, unfortunately we need to perform an API call to get ALL (very very bad). I need to know how many items we got to pick a random PetOfTheDay.
Once I got it we saved into the local stored with a date of today