refactor(app): Remove PetSate type and stored SortType.
We change the way we stored the pets. Instead on an object and a id by key we changed by an array of Pets because we have problems with the sort data cached.

To solve this BUG we created a sortType shared var and we improve getPetList that order the Pets[] also by sortType to prevent unnecessary API calls
We removed getPetsByIdRange && parsePetsResponse

We refactor PetListComponent to use this new methodology

Also we add in the dropdown sortBy = 'id' as default